update-schema

+++ Bunkering.Database/dbo/Tables/Payments.sql
@@ -20,13 +20,15 @@
     [LastRetryDate]      DATETIME2 (7)   NOT NULL,
     [Account]            NVARCHAR (MAX)  NOT NULL,
     [BankCode]           NVARCHAR (MAX)  NOT NULL,
-    [LateRenewalPenalty] DECIMAL (18, 2) NOT NULL,
-    [NonRenewalPenalty]  DECIMAL (18, 2) NOT NULL,
+    [LateRenewalPenalty] FLOAT (53)      NOT NULL,
+    [NonRenewalPenalty]  FLOAT (53)      NOT NULL,
     [Status]             NVARCHAR (MAX)  NOT NULL,
     CONSTRAINT [PK_Payments] PRIMARY KEY CLUSTERED ([Id] ASC),
     CONSTRAINT [FK_Payments_Applications_ApplicationId] FOREIGN KEY ([ApplicationId]) REFERENCES [dbo].[Applications] ([Id]),
     CONSTRAINT [FK_Payments_ExtraPayments_ExtraPaymentId] FOREIGN KEY ([ExtraPaymentId]) REFERENCES [dbo].[ExtraPayments] ([Id])
 );
+
+
